% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEST3EXP.R
\name{TEST3EXP}
\alias{TEST3EXP}
\title{Gives test statistic value and critical value for one parameter exponential distribution with group size 3}
\usage{
TEST3EXP(x1, x2, x3, alpha)
}
\arguments{
\item{x1}{is sample from Exp Population 1.}

\item{x2}{is sample from Exp Population 2.}

\item{x3}{is sample from Exp Population 3.}

\item{alpha}{is numeric value (level of significance)}
}
\description{
Gives test statistic value and critical value for one parameter exponential distribution with group size 3
}
